Modify the 'server' example to read PCM data from stdin, rather than a FIFO file.
This did a few cool things like:

  - Drop the CPU overhead by order of magnitude (from 20% to 1% on my old Mac Mini).
  - Remove the need completely for the InfiniteFifo class, since the scripts have been modified to never 'end'.
  - Allow a cool one-liner like:  "ffmpeg -f alse -i pulse -f wav -acodec pcm_s16le - | node server.js" to run a live podcast of your microphone.

NOTE: That this introduces the bug of static noise immediately after connecting, sometimes. I believe
the cause is that the PCM buffers from stdin are of unreliable length, and are sometimes beginning at
an offset that doesn't line up right with the 16-bit samples. This wan't happening with the PCMFIFO, since
it was using a bufferSize:1024, which "usually" would have the 16-bit samples lined up properly when a
new client connected. The fix should hopefully be coming in the next commit!